requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
strawberry-graphql>=0.209.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.1.0.post1
